refactor: change file name to lowercase
